Move checkout and form store branches to the SDK
