+++ web/app/connectors/_merlins-connector/checkout/commands.js
@@ -26,7 +26,7 @@ import {getCustomerEntityID, getCartBaseUrl} from '../selectors'
 import {receiveEntityID} from '../actions'
 import {PAYMENT_URL} from '../config'
 import {ADD_NEW_ADDRESS_FIELD} from '../../../containers/checkout-shipping/constants'
-import * as shippingSelectors from '../../../store/checkout/shipping/selectors'
+import * as shippingSelectors from 'progressive-web-sdk/dist/store/checkout/shipping/selectors'
 import {getCartItems} from 'progressive-web-sdk/dist/store/cart/selectors'
 import {getIsLoggedIn} from '../../../store/user/selectors'
 import {getShippingFormValues} from '../../../store/form/selectors'
