+++ web/app/connectors/_sfcc-connector/checkout/commands.js
@@ -11,7 +11,7 @@ import {parseShippingAddressFromBasket} from './parsers'
 import {getPaymentURL, getConfirmationURL} from '../config'
 import {receiveOrderConfirmationContents} from 'progressive-web-sdk/dist/integration-manager/results'
 import {getCardData} from 'progressive-web-sdk/dist/card-utils'
-import {getSelectedShippingMethodValue} from '../../../store/checkout/shipping/selectors'
+import {getSelectedShippingMethodValue} from 'progressive-web-sdk/dist/store/checkout/shipping/selectors'
 import {
     receiveShippingMethods,
     receiveShippingAddress,
