+++ web/app/containers/checkout-payment/actions.js
@@ -5,8 +5,8 @@
 import {browserHistory} from 'progressive-web-sdk/dist/routing'
 import {createAction} from 'progressive-web-sdk/dist/utils/action-creation'
 import {getPaymentBillingFormValues} from '../../store/form/selectors'
-import {getEmailAddress} from '../../store/checkout/selectors'
-import {getShippingAddress} from '../../store/checkout/shipping/selectors'
+import {getEmailAddress} from 'progressive-web-sdk/dist/store/checkout/selectors'
+import {getShippingAddress} from 'progressive-web-sdk/dist/store/checkout/shipping/selectors'
 import {submitPayment as submitPaymentCommand, initCheckoutPaymentPage} from 'progressive-web-sdk/dist/integration-manager/checkout/commands'
 import {splitFullName} from '../../utils/utils'
 import {handleCartExpiryError} from '../app/actions'
