+++ web/app/containers/checkout-shipping/partials/shipping-method.jsx
@@ -7,7 +7,7 @@ import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import * as ReduxForm from 'redux-form'
 
-import {getShippingMethods} from '../../../store/checkout/selectors'
+import {getShippingMethods} from 'progressive-web-sdk/dist/store/checkout/selectors'
 
 import Button from 'progressive-web-sdk/dist/components/button'
 import Field from 'progressive-web-sdk/dist/components/field'
