+++ web/app/store/checkout/selectors.js
@@ -1,38 +1,11 @@
-/* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
-/* Copyright (c) 2017 Mobify Research & Development Inc. All rights reserved. */
-/* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
 
-import Immutable from 'immutable'
 import {createSelector} from 'reselect'
-import {createGetSelector} from 'reselect-immutable-helpers'
 
-import {getFormValues} from '../form/selectors'
-import {getCheckout} from '../selectors'
+import {getSavedAddresses} from 'progressive-web-sdk/dist/store/checkout/shipping/selectors'
+import {getShippingSavedAddressID} from '../form/selectors'
 
-export const getEmailAddress = createGetSelector(getCheckout, 'email')
-
-export const getLocations = createGetSelector(getCheckout, 'locations', Immutable.Map())
-export const getCountries = createGetSelector(getLocations, 'countries', Immutable.List())
-export const getRegions = createGetSelector(getLocations, 'regions', Immutable.List())
-
-export const getBillingSameAsShipping = createGetSelector(getCheckout, 'billingSameAsShipping')
-
-export const getSelectedCountryID = (formKey) => createSelector(
-    getFormValues(formKey),
-    (values) => {
-        return values ? values.countryId : null
-    }
+export const getSelectedSavedShippingAddress = createSelector(
+    getSavedAddresses,
+    getShippingSavedAddressID,
+    (savedAddresses, selectedAddressID) => savedAddresses.find((address) => address.get('id') === selectedAddressID)
 )
-
-// Filter list of available regions based on what user has selected as country
-// Accepts a formKey that determines which form's country selection should be used
-export const getAvailableRegions = (formKey) => createSelector(
-    getRegions,
-    getSelectedCountryID(formKey),
-    (regions, id) => regions.filter((region) => region.get('countryId') === id)
-)
-
-export const getShippingMethods = createGetSelector(getCheckout, 'shippingMethods', Immutable.List())
-
-export const getCheckoutCustomContent = createGetSelector(getCheckout, 'custom')
-export const getLocationsCustomContent = createGetSelector(getLocations, 'custom')
