+++ web/app/store/selectors.js
@@ -5,7 +5,4 @@
 export const getUi = ({ui}) => ui
 export const getUser = ({user}) => user
 export const getCategories = ({categories}) => categories
-export const getCart = ({cart}) => cart
 export const getNotifications = ({notifications}) => notifications
-export const getCheckout = ({checkout}) => checkout
-export const getForm = ({form}) => form
